src/main/java/com/lapxpert/backend/chatbox/python/requirements.txt
# LapXpert AI Chat Dependencies
# Core dependencies for Vietnamese embedding model and GitHub AI Mistral Medium 3 (25.05) integration

# Database connectivity
psycopg2-binary==2.9.9

# Vietnamese text processing and embeddings
sentence-transformers==3.3.1
pyvi==0.1.1

# FastAPI and web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# AI Inference SDKs
azure-ai-inference==1.0.0b9  # GitHub AI (Mistral Medium 3) integration

# HTTP client for async operations (required by azure-ai-inference)
aiohttp==3.10.11

# Async database support
asyncpg==0.30.0

# Vector database support
pgvector==0.4.1

# Utilities
tqdm==4.67.1
numpy==2.2.1
torch==2.5.1

# Development and testing
pytest==8.3.4
httpx>=0.27.0,<0.28.0
